Refactor some of the JS
The click handler for the albums and artists was repeated pretty often
so I gave that its own function. Also the options for the draggable
objects was repeated a lot so that got a variable. I also removed an
unused function and added some spacing to the long lines of HTML.